Fix and refactor order delivery
